Create getting_sectors
This file goes on Fidelity's web page where it has overview of 11 sectors (https://eresearch.fidelity.com/eresearch/goto/markets_sectors/landing.jhtml). Using BeautifulSoup it opens the link to each sector and get's fundamentals for each sector and finally returns it in a pandas format for a quick analysis.